Put ConvergenceStrategy class into a separate file
